Move to struct.
