+++ crates/cache/src/lib.rs
@@ -1,8 +1,10 @@
 mod engine;
 mod helpers;
 mod items;
+mod reports;
 mod runfiles;
 
 pub use engine::CacheEngine;
 pub use helpers::*;
 pub use items::*;
+pub use reports::*;
